FAS-12 merge download.py and fastqheat.py (#16)
* Remove pycs

* Linters

* Format

* Configuration files

* Logging calls

* Shorten argparse code

* Remove commented-out code

* Minor cosmetic changes

* Remove default argparse actions

* More cosmetic changes

* Split handle_methods into different functions

* Remove else

* check for None implicitly

* Format files

* Remove unused code

* FAS-12 merge files

* Update download.py
